#24 refactor: seperate subscribe_add from subscribe
