+++ Ppap/app/src/main/java/com/jeongg/ppap/presentation/subscribe_add/SubscribeAddEvent.kt
@@ -1,4 +1,4 @@
-package com.jeongg.ppap.presentation.subscribe
+package com.jeongg.ppap.presentation.subscribe_add
 
 sealed class SubscribeAddEvent{
     data class EnteredTitle(val title: String): SubscribeAddEvent()
